[svn] Bumped version number.
Started adding -v (verbose) option. This needs to be implemented and also the command line parsing needs improving desperately.

Default font used is now crystal.ttf which is supplied with POV-Ray.

Updated example_config.txt to show how a font will be able to be specified in the next release.